requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
djangorestframework-simplejwt>=4.6,<5.0
psycopg2-binary>=2.8
djoser>=2.1,<3.0
drf-yasg>=1.20,<2.0
